+++ app/controllers/action_plans_controller.rb
@@ -89,7 +89,7 @@ class ActionPlansController < ApplicationController
           *columns.map { |column| project_action_plan_export_value(row, column, ownership) },
           *month_pairs.flat_map { |pair| [ row.public_send(pair[:target_column]), row.public_send(pair[:achievement_column]) ] },
           month_pairs.sum { |pair| row.public_send(pair[:target_column]).to_i },
-          month_pairs.sum { |pair| row.public_send(pair[:achievement_column]).to_i }
+          month_pairs.sum { |pair| row.public_send(pair[:achievement_column]).to_d }
         ]
       end
     end
